+++ src/main/java/edu/phema/elm_to_omop/translate/PhemaAssumptionException.java
@@ -1,7 +1,7 @@
-package edu.phema.elm_to_omop.model;
+package edu.phema.elm_to_omop.translate;
 
 public class PhemaAssumptionException extends Exception {
-  public PhemaAssumptionException(String message) {
-    super(message);
-  }
+    public PhemaAssumptionException(String message) {
+        super(message);
+    }
 }
