+++ src/main/java/edu/phema/elm_to_omop/translate/PhemaNotImplementedException.java
@@ -1,7 +1,7 @@
-package edu.phema.elm_to_omop.model;
+package edu.phema.elm_to_omop.translate;
 
 public class PhemaNotImplementedException extends Exception {
     public PhemaNotImplementedException(String message) {
-      super(message);
+        super(message);
     }
 }
